requirements.txt
streamlit
openai
tiktoken
python-dotenv